Restruture cdfg
